Fixed tests and moved methods for Monopoly
Fixed and confirmed unit tests, moved view methods to their own folder
